% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfplot.R
\name{acfplot}
\alias{acfplot}
\title{acf and pacf plot.
This is the acf and pacf plot function.}
\usage{
acfplot()
}
\arguments{
\item{x}{time series data}
}
\value{
the graphs of acf and pacf
}
\description{
acf and pacf plot.
This is the acf and pacf plot function.
}
\examples{
acfplot(timeseries)

}
