% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_series.R
\name{fred_series}
\alias{fred_series}
\title{FRED GNP CPI}
\usage{
fred_series(APIkey, series_id, observation_start)
}
\arguments{
\item{series_id}{This is a category of series}

\item{observation_start}{This is a date}
}
\value{

}
\description{
FRED GNP CPI
}
\examples{
fred_get_series(series_id="unrate")
}
